+++ tests/test_dfxp.py
@@ -183,6 +183,7 @@ class TestDFXPReader(ReaderTestingMixIn):
                 (HorizontalAlignmentEnum.START, VerticalAlignmentEnum.BOTTOM),
                 None,
                 None,
+                None,
             ),
             (
                 ((40, UnitEnum.PERCENT), (40, UnitEnum.PERCENT)),
@@ -191,12 +192,14 @@ class TestDFXPReader(ReaderTestingMixIn):
                 (HorizontalAlignmentEnum.START, VerticalAlignmentEnum.BOTTOM),
                 None,
                 None,
+                None,
             ),
             (
                 ((10, UnitEnum.PERCENT), (70, UnitEnum.PERCENT)),
                 None,
                 None,
                 (HorizontalAlignmentEnum.START, VerticalAlignmentEnum.BOTTOM),
+                None,
                 None,
                 None,
             ),