+++ tests/test_scc.py
@@ -173,12 +173,14 @@ class TestSCCReader(ReaderTestingMixIn):
                 (HorizontalAlignmentEnum.LEFT, VerticalAlignmentEnum.TOP),
                 None,
                 None,
+                None,
             ),
             (
                 ((10.0, UnitEnum.PERCENT), (83.0, UnitEnum.PERCENT)),
                 None,
                 None,
                 (HorizontalAlignmentEnum.LEFT, VerticalAlignmentEnum.TOP),
+                None,
                 None,
                 None,
             ),
@@ -199,6 +201,7 @@ class TestSCCReader(ReaderTestingMixIn):
                         (HorizontalAlignmentEnum.LEFT, VerticalAlignmentEnum.TOP),
                         None,
                         None,
+                        None,
                     )
                 ]
             },
@@ -209,6 +212,7 @@ class TestSCCReader(ReaderTestingMixIn):
                         None,
                         None,
                         (HorizontalAlignmentEnum.LEFT, VerticalAlignmentEnum.TOP),
+                        None,
                         None,
                         None,
                     )